remove gRPC support
