+++ src/python/shoalsoft/pants_opentelemetry_plugin/opentelemetry_config.py
@@ -1,0 +1,30 @@
+# Copyright (C) 2025 Shoal Software LLC. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Mapping
+
+
+@dataclass(frozen=True)
+class OtelParameters:
+    endpoint: str | None
+    certificate_file: str | None
+    client_key_file: str | None
+    client_certificate_file: str | None
+    headers: Mapping[str, str] | None
+    timeout: int | None
+    compression: str | None
+    insecure: bool | None
