+++ src/python/shoalsoft/pants_opentelemetry_plugin/requirements.txt
@@ -2,4 +2,3 @@
 opentelemetry-api==1.34.1
 opentelemetry-sdk==1.34.1
 opentelemetry-exporter-otlp==1.34.1
-grpcio==1.73.1
